new v1

+++ app/Http/Controllers/TaskController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\TodoTable;
+use Exception;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 
@@ -10,12 +11,22 @@ class TaskController extends Controller
 {
     public function getTasks()
     {
-        return response()->json(TodoTable::all());
+        // return response()->json(TodoTable::orderBy('id')->get());
+        $todo = TodoTable::orderBy('id')->get();
+        if ($todo) {
+            return response()->json($todo);
+        }
+        return response()->json(['status' => 'failed. No data available'], 404);
     }
 
     public function getTaskById($id)
     {
-        return response()->json(TodoTable::find($id));
+        // return response()->json(TodoTable::find($id));
+        $todo = TodoTable::orderBy('id')->find($id);
+        if ($todo) {
+            return response()->json($todo);
+        }
+        return response()->json(['status' => 'failed. No data available'], 404);
     }
 
     public function setTask(Request $request)
@@ -25,18 +36,23 @@ class TaskController extends Controller
     }
 
     public function updateTask(Request $request, $id)
-    {   
+    {
         $todo = TodoTable::findOrFail($id);
-        $todo->update($request->all());
-        $todo->save();
-        return response()->json([$todo, $request->all()], 201);
+        if ($todo) {
+            $todo->update($request->all());
+            $todo->save();
+            return response()->json($todo, 201);
+        }
+        return response(['status' => 'failed. No data found'], 404);
     }
 
     public function deleteTask($id)
     {
         $todo = TodoTable::find($id);
-        $todo->delete();
-
-        return response()->json(['status' => 'task with $id deleted'], 201);
+        if ($todo) {
+            $todo->delete();
+            return response()->json(['status' => 'task with ' . $id . ' deleted'], 201);
+        }
+        return response()->json(['status' => 'no task found'], 404);
     }
 }
